% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_alignment.R
\name{adjust_offset}
\alias{adjust_offset}
\title{adjust_offset}
\usage{
adjust_offset(
  img_paths,
  xy,
  t.subject,
  t.query,
  w = 100,
  s = 250,
  s2 = 0,
  verbose = F
)
}
\description{
adjust_offset
}
