% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_alignment.R
\name{cross_correlation}
\alias{cross_correlation}
\title{FFT Based Cross Correlation Computations}
\usage{
cross_correlation(img1, img2)
}
\description{
- http://www.learnpiv.org/fft/
  - https://chatgpt.com/share/674f5516-bb78-800f-9038-e00152429327
}
\details{
Define a function for cross-correlation
}
