% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_alignment.R
\name{img_alignment_offset}
\alias{img_alignment_offset}
\title{Compute image alignment offsets from FFT-based cross-correlation}
\usage{
img_alignment_offset(img1, img2, s = NULL, s2 = NULL, verbose = F)
}
\description{
Compute image alignment offsets from FFT-based cross-correlation
}
