% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcell2.examples.R
\name{list_examples}
\alias{list_examples}
\title{list_examples: List example notebooks from rcell2 example notebooks}
\usage{
list_examples()
}
\description{
Notebooks with several examples for different kinds of CellID data analyses.
}
\details{
Workflow examples are available as Rmd templates in their corresponding packages (rcell2, rcell2.cellid, and rcell2.magkc).
}
