% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles.R
\name{recenter.vector}
\alias{recenter.vector}
\title{Pone el origen de un vector en otro lado,
moviendo los elementos despues del borde al principio
y los de antes del borde quedan al final}
\usage{
recenter.vector(v, i)
}
\description{
Pone el origen de un vector en otro lado,
moviendo los elementos despues del borde al principio
y los de antes del borde quedan al final
}
