% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_alignment.R
\name{fftshift}
\alias{fftshift}
\title{A function to shift zero-frequency component to the center of the spectrum.}
\usage{
fftshift(x)
}
\description{
- http://matlab.izmiran.ru/help/techdoc/ref/fftshift.html
- https://stackoverflow.com/a/49863001
}
\details{
>  FFTSHIFT shifts the zero-frequency component to the center of the signal.
> In this case the signal is an image. A good visual guide is this.
> If you expand the original output image, you will see something akin to this:

Esto fue importante para poder encontrar consistentemente el máximo de la 
correlacion cruzada cerca del centro si el offset es pequeño.
}
