% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_alignment.R
\name{propagate_offsets}
\alias{propagate_offsets}
\title{propagate_offsets}
\usage{
propagate_offsets(
  img_paths,
  t.bud,
  xy,
  w = 50,
  s = NULL,
  s2 = NULL,
  xy_adj = F,
  verbose = F
)
}
\description{
propagate_offsets
}
